% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KostraDiverse.R
\name{AutoNetting}
\alias{AutoNetting}
\alias{AutoNettingNy}
\alias{AutoNettingIngenEndering}
\alias{AutoNettingCopy}
\title{AutoNetting}
\usage{
AutoNetting(
  nettinghierarki,
  hierarki,
  title = "nettinghierarki",
  toRemove = TRUE,
  singleLoop = FALSE
)

AutoNettingNy(
  nettinghierarki,
  hierarki,
  title = "nettinghierarki",
  loop = TRUE
)

AutoNettingIngenEndering(nettinghierarki, hierarki = NULL, title = NULL)

AutoNettingCopy(
  nettinghierarki,
  hierarki,
  title = "nettinghierarki",
  copyKode = "COPY"
)
}
\arguments{
\item{nettinghierarki}{nettinghierarki}

\item{hierarki}{hierarki}

\item{title}{title}
}
\value{
Ny versjon av nettinghierarki
}
\description{
Ulike varianter som ble utprøvd. AutoNettingCopy er løsningen som ble valgt.
}
\keyword{internal}
