% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanse_regnskap.R
\encoding{UTF8}
\name{balanse_regnskap}
\alias{balanse_regnskap}
\title{balanse_regnskap}
\usage{
balanse_regnskap(
  data,
  kapittelhierarki,
  kombinasjoner = NULL,
  regioner = NULL,
  storkombinasjoner = NULL,
  kapitler = NULL,
  omfang = NULL,
  output = "standard",
  printData = TRUE,
  lag0300 = FALSE,
  fixRegionkode = TRUE
)
}
\arguments{
\item{data}{inputdata}

\item{kapittelhierarki}{kapittelhierarki}

\item{kombinasjoner}{kombinasjoner kapitler og regnskapsomfang som skal med i output}

\item{regioner}{regioner som skal med i output. Kan også gjøre bruk av *, ?, ! og –}

\item{storkombinasjoner}{kombinasjoner kapitler,regnskapsomfang og region som skal med i output}

\item{kapitler}{kapitler som skal med i output}

\item{omfang}{regnskapsomfang som skal med i output}

\item{output}{kode for type output som (brukes av BalanseRegnskapBeregningHierarki og BalanseRegnskapBeregningInput)}

\item{printData}{Ved TRUE printes to første og to siste rader av alle inputdataene}

\item{lag0300}{Ved TRUE kopieres region 0301 til 0300 i inputdata}

\item{fixRegionkode}{Ved TRUE (default): Sørger for blanke i starten/slutten fjernes og at regionkoder får 4 eller 6 plasser og ledende nuller (gir warning ved endring av input)}
}
\value{
Data frame med samme variabler som data i input og med kapitler/regnskapsomfang i henhold til annen input
}
\description{
Funksjon som ligner på KostraRegnskap, men mye enklere.
}
\examples{

# Her brukes data som ligger i pakken der bare noen regioner er med.

inputdata <- kr_data("balansedata")  
hierarki <- kr_data("kapittelhierarki")  

z <- balanse_regnskap(inputdata, hierarki, regioner = c("2021", "2022", "0301"))

}
\seealso{
\code{\link{KostraRegnskap}}, \code{\link{HierarchyCompute}}, \code{\link{HierarchicalWildcardGlobbing}}
}
