% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kostra_regnskap.R
\name{beredt}
\alias{beredt}
\title{Forbehandling av input}
\usage{
beredt(...)
}
\arguments{
\item{...}{Input til \code{\link{KostraRegnskapEnPeriode}}}
}
\value{
Liste med data som er forberedt
}
\description{
Forbehandling av input
}
