% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatt_hierarki.R
\name{ett_flatt_hierarki}
\alias{ett_flatt_hierarki}
\title{Ett flatt artshierarki}
\usage{
ett_flatt_hierarki(
  artshierarki,
  artshierarki_nettinger,
  artshierarki_nettinger_kasse,
  periode = NULL,
  fix = TRUE,
  flatt = TRUE,
  g = 0
)
}
\arguments{
\item{artshierarki}{artshierarki}

\item{artshierarki_nettinger}{artshierarki_nettinger}

\item{artshierarki_nettinger_kasse}{artshierarki_nettinger_kasse}

\item{periode}{Ved flere perioder i hierarkier, må en periode velges.}

\item{fix}{Ved `TRUE` (default): Som `fix` i \code{\link{flatt_hierarki}}
samt utvidelser basert på from-koden `"COPY"`}

\item{flatt}{Hvorvidt flatt hierarki skal returneres. 
Ved `FALSE` returneres resultat etter `fix` og
kombinering til ett hierarki.}

\item{g}{Flatt hierarki begrenses til `abs(sign)>g`.}
}
\value{
Ett (flatt) hierarki kodet på spesiell måte slik at alt samles.
}
\description{
artshierarki, artshierarki_nettinger og artshierarki_nettinger_kasse 
settes sammen til ett hierarki. Deretter genereres flatt hierarki fra dette.
}
\details{
Underliggende funksjon er samme funksjon som brukes til beregningsteter. 
Et unntak er at denne funksjonen ikke håndterer formler.
}
\examples{
kdata <- kr_data("kostraRegnskapDataPen") 
funksjonshierarki <- kdata$funksjonshierarki 
artshierarki <- kdata$artshierarki          
inputdata_saer <- kdata$data_saer
artshierarki_nettinger <- kdata$artshierarki_nettinger
artshierarki_nettinger_kasse <- kdata$artshierarki_nettinger_kasse

head(ett_flatt_hierarki(artshierarki, artshierarki_nettinger, 
             artshierarki_nettinger_kasse, periode = "2015", g = 1))
head(ett_flatt_hierarki(artshierarki, artshierarki_nettinger, 
             artshierarki_nettinger_kasse, periode = "2015", flatt = FALSE))
ett_flatt_hierarki(artshierarki, artshierarki_nettinger, 
             artshierarki_nettinger_kasse, periode = "2016", g = 2)
}
\seealso{
\code{\link{flatt_hierarki}}
}
