% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatt_hierarki.R
\name{flatt_hierarki}
\alias{flatt_hierarki}
\title{Flatt hierarki}
\usage{
flatt_hierarki(
  hierarki,
  copy_hierarki = NULL,
  periode = NULL,
  fix = TRUE,
  flatt = TRUE,
  g = 0
)
}
\arguments{
\item{hierarki}{hierarki}

\item{copy_hierarki}{Ekstra hierarki som skal brukes når \code{hierarki} er et nettinghierarki som skal utvides basert på from-koden \code{"COPY"}.}

\item{periode}{Ved flere perioder i \code{hierarki}, må en periode velges.}

\item{fix}{Ved \code{TRUE} (default): Sørger for blanke i starten/slutten fjernes og at koder som er tall får 3 plasser og ledende nuller (gir warning ved endring av input).}

\item{flatt}{Hvorvidt flatt hierarki skal returneres. Ved \code{FALSE} returneres resultat etter \code{fix} og bruk av \code{copy_hierarki}.}

\item{g}{Flatt hierarki begrenses til \code{abs(sign)>g}.}
}
\value{
Et (flatt) hierarki
}
\description{
Omkoder hierarki slik at alle from-koder er koder fra input-data
}
\examples{
kdata <- kr_data("kostraRegnskapDataPen")
funksjonshierarki <- kdata$funksjonshierarki
artshierarki <- kdata$artshierarki          
inputdata_saer <- kdata$data_saer
artshierarki_nettinger <- kdata$artshierarki_nettinger
artshierarki_nettinger_kasse <- kdata$artshierarki_nettinger_kasse

head(flatt_hierarki(funksjonshierarki, periode = "2015"))
flatt_hierarki(artshierarki, periode = "2016", g = 2)
flatt_hierarki(artshierarki_nettinger, artshierarki, periode = "2016", g = 1)
head(flatt_hierarki(artshierarki_nettinger_kasse, artshierarki, periode = "2016", flatt = FALSE))
}
\seealso{
\code{\link{ett_flatt_hierarki}}
}
