% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kostra_regnskap.R
\name{formel_korreksjon}
\alias{formel_korreksjon}
\title{Formel-korreksjon av sett av matrisekomponenter}
\usage{
formel_korreksjon(a, af, fm)
}
\arguments{
\item{a}{Sett av matrisekomponenter}

\item{af}{Korresponderende  sett av matrisekomponenter med kombinasjoner som trengs for formler}

\item{fm}{Output fra \code{FormelMatrise}  (gammel intern funksjon for beregningstester)}
}
\value{
Korrigert versjon av input \code{a}
}
\description{
Formel-korreksjon av sett av matrisekomponenter
}
