% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kostra_regnskap.R
\name{get_a1234}
\alias{get_a1234}
\title{Fire sett av matrisekomponenter}
\usage{
get_a1234(
  data,
  data_saer = NULL,
  funksjonshierarki,
  artshierarki,
  arts32 = NULL,
  arts41 = NULL,
  regioner,
  kombinasjoner,
  rowsInputArt,
  handleDuplicated,
  useMatrixToDataFrame = TRUE,
  ...
)
}
\arguments{
\item{data}{data}

\item{data_saer}{data_saer}

\item{funksjonshierarki}{funksjonshierarki}

\item{artshierarki}{artshierarki}

\item{arts32}{artshierarki_nettinger omgjort til hierarki som kan beregnes fra \code{data_saer}}

\item{arts41}{artshierarki_nettinger_kasse omgjort til hierarki som kan beregnes fra \code{data}}

\item{regioner}{regioner}

\item{kombinasjoner}{kombinasjoner}

\item{rowsInputArt}{Denne ble beregnet som \code{kombinasjoner$art \%in\% unique(c(unique(data$art), unique(data_saer$art)))}.
Kan sannsynligvis fjernes som parameter og heller beregnes på nytt.
Brukes for å ta bort koder for å hindre fordobling fra både hierarki og nettinghierarki.}

\item{handleDuplicated}{Denne bør være \code{stop} når ved beregning av id-bidrag siden rad-nummer ikke kan summeres.}

\item{useMatrixToDataFrame}{useMatrixToDataFrame}

\item{...}{Ubrukte parametere}
}
\value{
Fire sett \code{HierarchyCompute}-output ved bruk av \code{output = "matrixComponents"}
}
\description{
Med input fra \code{\link{beredt}}-output
}
