% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_bidrag.R
\name{id_bidrag}
\alias{id_bidrag}
\alias{fun_belop}
\title{Tekststreng som angir bidrag til \code{sum(sign*belop)}}
\usage{
id_bidrag(ind, sign, data, id = "id", belop = "belop")

fun_belop(ind, sign, data, id = "id", belop = "belop")
}
\arguments{
\item{ind}{Radnummer i data}

\item{sign}{Faktor data skal multipliseres med. Samme lengde som \code{ind}.}

\item{data}{En data frame}

\item{id}{Variabel i data med id}

\item{belop}{Variabel i data med belop}
}
\value{
En tekststreng. Dvs.lengde-1-character-variabel.
}
\description{
Mer presist: Bidrag til \code{sum(sign[ind > 0] * data[["belop"]][ind])},
der \code{ind} er vektor med radnummer i data og
der  \code{sign} har samme lengde som \code{ind}.
Funksjonen som gir det numeriske resultatbeløpet er med (\code{fun_belop}).
}
\examples{
# Genererer input
df <- data.frame(id = paste0("id", 1:9), belop = 11 * 1:9)
df
ind <- c(8, 5, 7)
sign <- c(1, -1, 2)

# Den aktuelle summen er 187
sum(sign * df[["belop"]][ind])

# Bidrag til denne summen
id_bidrag(ind, sign, df)

# Sum med to første indekser
sum(c(1, -1) * df[["belop"]][c(8, 5)])

# Bidrag til denne summen, nå med '+' og '-'
id_bidrag(c(8, 5), c("+", "-"), df)

# 0-indekser som ignoreres kan være med
ind2 <- c(8, 0, 5, 0, 7)
sign2 <- c(1, -1, -1, -3, 2)

# Summen beregnes generelt slik
sum(sign2[ind2 > 0] * df[["belop"]][ind2])

# Samme med funksjonen fun_belop 
fun_belop(ind2, sign2, df)

# 0-indekser ignoreres av id_bidrag
id_bidrag(ind2, sign2, df)
}
