% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_bidrag_matrix.R
\name{id_bidrag_matrix}
\alias{id_bidrag_matrix}
\title{\code{id_bidrag} med multippel \code{ind}-parameter og multippel \code{sign}-parameter}
\usage{
id_bidrag_matrix(sign_matrix, ind_matrix, data, fun_id_bidrag = id_bidrag)
}
\arguments{
\item{sign_matrix}{Matrise der ikke-0 i hver kolonne gir opphav til \code{sign}-
og \code{ind}-input til \code{id_bidrag_vector}.
Det er fordelaktig om denne matrisa er av sparse type siden den kommer til å bli konvertert til en slik.}

\item{ind_matrix}{Input til \code{id_bidrag_vector}.
Denne matrisa kan være sparse, men den vil internt bli konvertert til en vanlig matrise.}

\item{data}{\code{data}-parameter til \code{id_bidrag}}

\item{fun_id_bidrag}{Funksjonen \code{id_bidrag}. Kan endres.}
}
\value{
En vanlig matrise
}
\description{
\code{\link{id_bidrag}} med multippel \code{ind}-parameter i matrise
og multippel \code{sign}-parameter i matrise slik at output er en matrise.
Også mulig med \code{\link{fun_belop}} som alternativ input-funksjon slik at output blir
resultatbeløp som en numerisk matrise.
Beregningene går via \code{\link{id_bidrag_vector}}.
}
\examples{
# Genererer input

df <- data.frame(id = paste0("id", 1:9), belop = 11 * 1:9)
df

ind_matrix <- matrix(0, 7, 5)
ind_matrix[2:4, 1] <- c(8, 5, 7)
ind_matrix[5:7, 2] <- c(8, 5, 7)
ind_matrix[, 3] <- 1:7
ind_matrix[c(1, 7), 5] <- 9
colnames(ind_matrix) <- paste0("im", 1:5)
ind_matrix

sign_matrix <- matrix(0, 7, 6)
sign_matrix[2:4, 1] <- c(1, -1, 2)
sign_matrix[5:7, 2] <- c(1, -1, 2)
sign_matrix[c(1, 7), 3] <- 1:2
sign_matrix[7, 4] <- -1
sign_matrix[, 6] <- rep(1, 7)
colnames(sign_matrix) <- paste0("sm", 1:6)
sign_matrix

# Sparse versjon av sign_matrix
sign_Matrix <- Matrix::Matrix(sign_matrix)
sign_Matrix

# Output med fun_belop og id_bidrag (denne printes ikke)
id_bidrag_matrix(sign_matrix, ind_matrix, df, fun_id_bidrag = fun_belop)
out <- id_bidrag_matrix(sign_matrix, ind_matrix, df)

# Med mindre matriser som input (en av dem er sparse)
id_bidrag_matrix(sign_matrix[2:3, c(1, 6)], ind_matrix[4:5, c(1, 3)], df, fun_belop)
id_bidrag_matrix(sign_Matrix[2:3, c(1, 6)], ind_matrix[4:5, c(1, 3)], df)

}
\keyword{internal}
