% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_bidrag_vector.R
\name{id_bidrag_vector}
\alias{id_bidrag_vector}
\alias{belop_vector}
\title{\code{id_bidrag} med multippel \code{ind}-parameter}
\usage{
id_bidrag_vector(sign, ind, ind_matrix, data, fun_id_bidrag = id_bidrag)

belop_vector(..., fun_id_bidrag = fun_belop)
}
\arguments{
\item{sign}{Faktor data skal multipliseres med. Samme lengde som \code{ind} nedenfor}

\item{ind}{Radnummer i \code{ind_matrix} som skal brukes}

\item{ind_matrix}{Matrise slik at hver kolonne i \code{ind_matrix[ind, ]} kan brukes som \code{ind}-parameter til \code{id_bidrag} sammen med \code{sign}}

\item{data}{\code{data}-parameter til \code{id_bidrag}}

\item{fun_id_bidrag}{Funksjonen \code{id_bidrag}. Kan endres.}

\item{...}{Parametere som videresendes}
}
\value{
Vektor med multippel \code{id_bidrag}-output
}
\description{
\code{\link{id_bidrag}} med multippel \code{ind}-parameter i matrise slik at output er en vektor.
Samt en tilsvarende funksjon som gir resultatbeløp som en numerisk vektor.
}
\examples{
# Genererer input
df <- data.frame(id = paste0("id", 1:9), belop = 11 * 1:9)
df
ind_matrix <- matrix(0, 7, 5)
ind_matrix[2:4, 1] <- c(8, 5, 7)
ind_matrix[5:7, 2] <- c(8, 5, 7)
ind_matrix[, 3] <- 1:7
ind_matrix[c(1, 7), 5] <- 9
ind_matrix
sign <- c(1, -1, 2)

# Her er 1. element som i 1. id_bidrag-eksempel
id_bidrag_vector(sign, 2:4, ind_matrix, df)
belop_vector(sign, 2:4, ind_matrix, df)

# Her er 2. element som i 1. id_bidrag-eksempel
id_bidrag_vector(sign, 5:7, ind_matrix, df)
belop_vector(sign, 5:7, ind_matrix, df)

id_bidrag_vector(1:2, c(1, 7), ind_matrix, df)
belop_vector(1:2, c(1, 7), ind_matrix, df)

# Også mulig med bare en rad i matrisa
id_bidrag_vector(-1, 7, ind_matrix, df)
belop_vector(-1, 7, ind_matrix, df)

# Faktisk også mulig med ingen rader
id_bidrag_vector(integer(0), integer(0), ind_matrix, df)
belop_vector(integer(0), integer(0), ind_matrix, df)
}
\keyword{internal}
