% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr_data.R
\docType{data}
\name{kostraRegnskapData}
\alias{kostraRegnskapData}
\alias{kostraRegnskapDataPen}
\alias{balanseRegnskapData}
\alias{balanseRegnskapDataPen}
\title{Datasets returned by kr_data()}
\description{
Datasets returned by kr_data()
}
\keyword{datasets}
\keyword{internal}
