% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_generate.R
\name{uuid_generate}
\alias{uuid_generate}
\alias{uuid_generate_time}
\alias{uuid_generate_random}
\title{\code{UUIDgenerate} with \code{use.time} specified}
\usage{
uuid_generate(n)

uuid_generate_time(n)

uuid_generate_random(n)
}
\arguments{
\item{n}{number of UUIDs to generate}
}
\value{
character vector
}
\description{
\code{\link{UUIDgenerate}} from package \code{uuid} with \code{use.time} specified
\itemize{
\item \strong{\code{uuid_generate}}  uses default \code{use.time}
\item \strong{\code{uuid_generate_time}}  uses \code{use.time = TRUE}
\item \strong{\code{uuid_generate_random}}  uses \code{use.time = FALSE}
}
}
\examples{
uuid_generate_time(1)
uuid_generate_time(2)
uuid_generate_random(3)
}
