% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_generate.R
\name{uuid_generate_time}
\alias{uuid_generate_time}
\title{\code{UUIDgenerate} med \code{use.time = TRUE}}
\usage{
uuid_generate_time(n)
}
\arguments{
\item{n}{number of UUIDs to generate}
}
\value{
id-er
}
\description{
\code{\link{UUIDgenerate}} fra pakken \code{uuid} med \code{use.time = TRUE}
}
\examples{
uuid_generate_time(2)
}
