% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanse_regnskap.R
\encoding{UTF8}
\name{balanse_regnskap}
\alias{balanse_regnskap}
\title{balanse_regnskap}
\usage{
balanse_regnskap(
  data,
  kapittelhierarki,
  kombinasjoner = NULL,
  regioner = NULL,
  storkombinasjoner = NULL,
  kapitler = NULL,
  omfang = NULL,
  output = "standard",
  printData = TRUE,
  lag0300 = FALSE,
  fixRegionkode = TRUE,
  bidrag = TRUE,
  generer_id = TRUE,
  bidrag_var = "source",
  id_var = "UUID",
  fun_id_bidrag = id_bidrag,
  fun_generer_id = uuid_generate_time
)
}
\arguments{
\item{data}{inputdata}

\item{kapittelhierarki}{kapittelhierarki}

\item{kombinasjoner}{kombinasjoner kapitler og regnskapsomfang som skal med i output}

\item{regioner}{regioner som skal med i output. Kan også gjøre bruk av *, ?, ! og –}

\item{storkombinasjoner}{kombinasjoner kapitler,regnskapsomfang og region som skal med i output}

\item{kapitler}{kapitler som skal med i output}

\item{omfang}{regnskapsomfang som skal med i output}

\item{output}{kode for type output som (brukes av BalanseRegnskapBeregningHierarki og BalanseRegnskapBeregningInput)}

\item{printData}{Ved TRUE printes to første og to siste rader av alle inputdataene}

\item{lag0300}{Ved TRUE kopieres region 0301 til 0300 i inputdata}

\item{fixRegionkode}{Ved TRUE (default): Sørger for blanke i starten/slutten fjernes og at regionkoder får 4 eller 6 plasser og ledende nuller (gir warning ved endring av input)}

\item{bidrag}{Ved \code{TRUE} vil det genereres tekststreng
med funksjonen spesifisert med parameteren \code{fun_id_bidrag}.}

\item{generer_id}{Ved \code{TRUE} vil det genereres nye id-er i output-data
med funksjonen spesifisert med parameteren \code{fun_generer_id}.}

\item{bidrag_var}{Navn på variabel i output som genereres når \code{bidrag = TRUE}.}

\item{id_var}{Navn på id-variabel i input/output.}

\item{fun_id_bidrag}{\code{\link{id_bidrag}} eller en tilsvarende funksjon.}

\item{fun_generer_id}{\code{\link{uuid_generate_time}} eller en tilsvarende funksjon.}
}
\value{
Data frame med samme variabler som data i input og med kapitler/regnskapsomfang i henhold til annen input
}
\description{
Funksjon som ligner på KostraRegnskap, men mye enklere.
}
\examples{

# Her brukes data som ligger i pakken der bare noen regioner er med.

inputdata <- kr_data("balansedata")  
hierarki <- kr_data("kapittelhierarki")  

z <- balanse_regnskap(inputdata, hierarki, regioner = c("2021", "2022", "0301"))

}
\seealso{
\code{\link{KostraRegnskap}}, \code{\link{HierarchyCompute}}, \code{\link{HierarchicalWildcardGlobbing}}
}
