% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kostra_regnskap.R
\name{kostra_regnskap}
\alias{kostra_regnskap}
\title{kostra_regnskap}
\usage{
kostra_regnskap(..., output = "en")
}
\arguments{
\item{...}{dots}

\item{output}{output}
}
\value{
output
}
\description{
kostra_regnskap
}
