% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kostra_regnskap.R
\name{kostra_regnskap}
\alias{kostra_regnskap}
\title{kostra_regnskap}
\usage{
kostra_regnskap(
  data,
  funksjonshierarki,
  artshierarki,
  data_saer = NULL,
  artshierarki_nettinger = NULL,
  artshierarki_nettinger_kasse = NULL,
  ...,
  output = "en",
  bidrag = FALSE
)
}
\arguments{
\item{...}{dots}

\item{output}{output}
}
\value{
output
}
\description{
kostra_regnskap
}
