% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr_data.R
\encoding{UTF8}
\name{kr_data}
\alias{kr_data}
\title{Return a data set}
\usage{
kr_data(dataset, path = NULL)
}
\arguments{
\item{dataset}{Name of data set within the kostraregnskap package}

\item{path}{When non-NULL the data set is read from "path/dataset.RData"}
}
\value{
The data set
}
\description{
Return a data set
}
\examples{
 z  <- kr_data("balanseRegnskapDataPen")
 z  <- kr_data("balanseRegnskapData")
 z  <- kr_data("kostraRegnskapDataPen")
 z  <- kr_data("kostraRegnskapData")

}
